refactor(sdk): fix typo in test file name
