+++ packages/rs-sdk/tests/fetch/mod.rs
@@ -11,7 +11,7 @@ mod common;
 mod config;
 mod contested_resource;
 mod contested_resource_identity_votes;
-mod contested_resource_pools_by_ts;
+mod contested_resource_polls_by_ts;
 mod contested_resource_vote_state;
 mod contested_resource_voters;
 mod data_contract;
